Increase alignment of vector tables
While VBAR needs only 11 bit alignment, here we fix a 12 bit alignment.
This means that each vector table takes an entire 4k page and so the
bottom 12 bits are completely determined.

This means that aliases to the vector table do not need to worry about
offsets into the page since those lower 12 bits are unaffected.