Fix icache invalidation
This is a fun one!
ARMv8 requires all _data_ caches to behave like PIPT caches.
But this is not true of instruction caches.  They can also behave
like VIPT or other virtually-indexed things.

This has the direct consequence that if a PA has multiple aliases then
a DC should affect all aliases, but an IC does not have to. Since here
we use an alias for writing to the vector table, we have to ensure
that we perform IC operations only over the final VA (in this case,
identically mapped to the PA) stored in VBAR.