(lib/argparse) fix nullptr deref on NULL shortname
If the short or long names were NULL, then 1+short_name is passed to
the various functions, and the following line triggered UB:
> if (short_name && strcmp(s, short_name))

So, now pass short_name as-is, and check for NULL in the functions.

Perhaps - and -- should not appear in the OPT/FLAGs at all, but be added
by the library later.