(litmus_tests) add SEA_R_detect test
This test can detect if the device exhibits synchronous external aborts
on a load.

By attempting to write to unbacked memory (i.e. no physical device there)
and then catching the exception, if we saw a Synchronous SError then we
know that the device must support them.

WARNING: If the device returns an imprecise asynchronous error on such
an event, then the whole harness is toast, so this is not safe to batch.