(litmus/var_infos) fix: check for at least one var
It's important for tests to have at least one var
which owns some virtual address space itself.

Instead of listing them explicitly one-by-one,
use a catch-all `var_owns_region`.

Note that tests that have no pins or offset vars
might not need this check at all, so we may be
able to relax it in future.