(valloc/arenas) add arena allocation
Allocation of test run data with arenas reduces the contention on the
allocator alloclist, which would otherwise be hit by many very small
allocations which all need to be allocated and free'd together.

This lets us allocate a whole batch of per-thread run data at once,
and then free it all together at the end.