(Makefile) propagate -Wextra everywhere
Disables -Wunused-parameter and -Wsign-compare.

The first is very noisy, and unhelpful.
There are many callbacks,
and condtionally-compiled use of parameters (e.g. for debugging)
which makes it too noisy to use productively.

The second is noisy due to mixing of int and u64 as loop comparators,
it's temporarily disabled as overflow of those integers seems improbable.
For completeness we should enable it and fix all the places we could have
a signed integer overflow.