add affinity control with taskset
When there are more than 4 cores, or mixed architecture (e.g. big.LITTLE),
then we want to be able to specify which we run on.

This makes `make kvm` produce a shell script that runs taskset on qemu.
By default QEMU is pinned to cores 0-7.
Can specify which cores with e.g.:
> AFFINITY=0xf ./kvm_litmus ...

Or compile a kvm_litmus without affinity control by
> make kvm AFFINITY=