add safety checks for big.LITTLE cores
Detect the case where we run on heterogenous machines,
e.g. arm big.LITTLE cores and crash angrily if we detect different
implementations on the hardware threads.

Also as a sanity check make sure we don't migrate from big/little cores
to the other. One hopes this really never happens, but the test results
would be invalid if it did, so if we catch the OS red-handed then we
similarly get angry and crash.